权限查询

+++ darkblue-application/src/main/java/org/dblue/application/module/permission/application/service/impl/PermissionApplicationServiceImpl.java
@@ -18,8 +18,17 @@ package org.dblue.application.module.permission.application.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.dblue.application.module.permission.application.dto.PermissionQueryDto;
 import org.dblue.application.module.permission.application.service.PermissionApplicationService;
+import org.dblue.application.module.permission.application.vo.PermissionVo;
+import org.dblue.application.module.permission.domain.PermissionDomainQueryService;
 import org.dblue.application.module.permission.domain.service.PermissionDomainService;
+import org.dblue.application.module.permission.errors.PermissionErrors;
+import org.dblue.application.module.permission.infrastructure.entiry.Permission;
+import org.dblue.application.module.role.infrastructure.domain.service.RolePermissionDomainService;
+import org.dblue.common.exception.ServiceException;
+import org.springframework.beans.BeanUtils;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,6 +44,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class PermissionApplicationServiceImpl implements PermissionApplicationService {
 
     private final PermissionDomainService permissionDomainService;
+    private final RolePermissionDomainService rolePermissionDomainService;
+    private final PermissionDomainQueryService permissionDomainQueryService;
     /**
      * 权限删除
      *
@@ -43,6 +54,32 @@ public class PermissionApplicationServiceImpl implements PermissionApplicationSe
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void delete(String id) {
+        long count = rolePermissionDomainService.countByPermissionId(id);
+        if(count > 0){
+            throw new ServiceException(PermissionErrors.PERMISSION_ROLE_EXITS);
+        }
         permissionDomainService.delete(id);
+    }
+
+    /**
+     * 分页查询权限信息
+     *
+     * @param query 查询条件
+     * @return 权限列表
+     */
+    @Override
+    public Page<PermissionVo> findByPage(PermissionQueryDto query) {
+        Page<Permission> page = permissionDomainQueryService.findByPage(query);
+        if(page.isEmpty()){
+
+            return Page.empty(query.toJpaPage());
+        }
+        return page.map(this::build);
+    }
+
+    private PermissionVo build(Permission permission) {
+        PermissionVo permissionVo = new PermissionVo();
+        BeanUtils.copyProperties(permission,permissionVo);
+        return  permissionVo;
     }
 }
